requirements.txt
alabaster==0.7.8
appdirs==1.4.3
apturl==0.5.2
asgiref==3.5.2
autopep8==1.6.0
Babel==2.6.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2022.6.15
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.5
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==1.0.0
Django==4.0.6
django-debug-toolbar==3.2.1
django-filter==22.1
django-mysql==4.7.1
django-rest-framework-nested==0.0.1
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.1.0
dnspython==1.16.0
docopt==0.6.2
docutils==0.16
drf-nested-routers==0.93.4
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
feedparser==5.2.1
filelock==3.8.0
future==0.18.2
gitsome==0.8.0
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
itypes==1.2.0
Jinja2==2.10.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
mysql-client==0.0.1
mysqlclient==2.0.3
netifaces==0.10.4
numpydoc==0.7.0
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pipenv==2022.8.19
platformdirs==2.5.2
ply==3.11
prompt-toolkit==2.0.10
protobuf==3.6.1
proton-client==0.7.1
protonvpn-cli==3.11.1
protonvpn-gui==1.7.0
protonvpn-nm-lib==3.8.0
psutil==5.5.1
pycairo==1.16.2
pycodestyle==2.8.0
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==2.4.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gnupg==0.4.5
python3-openid==3.2.0
pythondialog==3.4.0
pytz==2021.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
roman==2.0.0
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
Sphinx==1.8.5
sqlparse==0.4.2
systemd-python==234
toml==0.10.2
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.16.3
virtualenv-clone==0.5.7
wadllib==1.3.3
wcwidth==0.1.8
whitenoise==6.2.0
xkit==0.0.0
xonsh==0.9.13
zipp==1.0.0
